monitor: Move PolicyMatchType into pkg/monitor/api
This type is used by external components. Move it into pkg/monitor/api
to reduce additional dependencies being pulled in.
